Refactor bauble methods into abstract class (#3)
* Refactor some bauble methods into an abstract class

* This was needed to stop TC4 from deploading the OG version

* remove unneeded clientonly

* Add note
